+++ cmd/mikrotik-hosts-parser/main.go
@@ -6,7 +6,8 @@ import (
 	"path/filepath"
 
 	"github.com/fatih/color"
-	"github.com/tarampampam/mikrotik-hosts-parser/v4/internal/pkg/cli"
+
+	"gh.tarampamp.am/mikrotik-hosts-parser/v4/internal/pkg/cli"
 )
 
 // exitFn is a function for application exiting.