+++ internal/pkg/cli/serve/flags.go
@@ -12,7 +12,7 @@ import (
 	"github.com/go-redis/redis/v8"
 	"github.com/spf13/pflag"
 
-	"github.com/tarampampam/mikrotik-hosts-parser/v4/internal/pkg/env"
+	"gh.tarampamp.am/mikrotik-hosts-parser/v4/internal/pkg/env"
 )
 
 type flags struct {