+++ internal/pkg/http/handlers/api/settings/handler.go
@@ -5,9 +5,8 @@ import (
 	"encoding/json"
 	"net/http"
 
-	"github.com/tarampampam/mikrotik-hosts-parser/v4/internal/pkg/cache"
-
-	"github.com/tarampampam/mikrotik-hosts-parser/v4/internal/pkg/config"
+	"gh.tarampamp.am/mikrotik-hosts-parser/v4/internal/pkg/cache"
+	"gh.tarampamp.am/mikrotik-hosts-parser/v4/internal/pkg/config"
 )
 
 type (